+++ tests/cpp/name_transforms_test.cc
@@ -21,7 +21,7 @@
 
 #include <gtest/gtest.h>
 #include <tvm/runtime/container/string.h>
-#include <tvm/runtime/name_mangling.h>
+#include <tvm/runtime/name_transforms.h>
 
 using namespace tvm::relay::backend;
 using namespace tvm::runtime;
